move old config types to only point of use: tests
