+++ hack/import-restrictions.json
@@ -32,6 +32,20 @@
     "checkedPackageRoots": [
       "github.com/openshift/origin/pkg"
     ],
+    "ignoredSubTrees": [],
+    "forbiddenImportPackageRoots": [
+      "github.com/openshift/origin/test"
+    ],
+    "allowedImportPackageRoots": [
+      "vendor",
+      ""
+    ]
+  },
+
+  {
+    "checkedPackageRoots": [
+      "github.com/openshift/origin/pkg"
+    ],
     "ignoredSubTrees": [
       "github.com/openshift/origin/pkg/apiserver/admission",
       "github.com/openshift/origin/pkg/autoscaling/admission/clusterresourceoverride",
@@ -140,24 +154,7 @@
     ],
     "ignoredSubTrees": [
       "github.com/openshift/origin/pkg/cmd/openshift-integrated-oauth-server",
-      "github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/openshiftkubeapiserver",
-      "github.com/openshift/origin/pkg/cmd/server/origin",
-      "github.com/openshift/origin/pkg/cmd/server/apis/config/validation",
-      "github.com/openshift/origin/pkg/oc/cli/admin",
-      "github.com/openshift/origin/pkg/oc/cli/admin/cert",
-      "github.com/openshift/origin/pkg/oc/cli/admin/createclient",
-      "github.com/openshift/origin/pkg/oc/cli/admin/createerrortemplate",
-      "github.com/openshift/origin/pkg/oc/cli/admin/createkubeconfig",
-      "github.com/openshift/origin/pkg/oc/cli/admin/createlogintemplate",
-      "github.com/openshift/origin/pkg/oc/cli/admin/createproviderselectiontemplate",
-      "github.com/openshift/origin/pkg/oc/cli/admin/groups/sync",
-      "github.com/openshift/origin/pkg/cmd/openshift-osinserver",
-      "github.com/openshift/origin/pkg/oc/lib/groupsync",
-      "github.com/openshift/origin/pkg/oc/lib/groupsync/ad",
-      "github.com/openshift/origin/pkg/oc/lib/groupsync/groupdetector",
-      "github.com/openshift/origin/pkg/oc/lib/groupsync/rfc2307",
-      "github.com/openshift/origin/pkg/oc/lib/groupsync/syncerror",
-      "github.com/openshift/origin/pkg/oc/lib/signercertoptions"
+      "github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/openshiftkubeapiserver"
     ],
     "forbiddenImportPackageRoots": [
       "vendor/github.com/openshift/oauth-server"
