+++ test/integration/oauth_external_test.go
@@ -22,11 +22,11 @@ import (
 
 	"github.com/openshift/oc/pkg/helpers/tokencmd"
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
 	oauthutil "github.com/openshift/origin/pkg/oauth/util"
 	userclient "github.com/openshift/origin/pkg/user/generated/internalclientset/typed/user/internalversion"
 	testutil "github.com/openshift/origin/test/util"
 	testserver "github.com/openshift/origin/test/util/server"
+	configapi "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
 )
 
 // TestWebhookTokenAuthn checks Tokens directly against an external
