+++ test/integration/oauth_htpasswd_test.go
@@ -9,10 +9,10 @@ import (
 	restclient "k8s.io/client-go/rest"
 
 	"github.com/openshift/oc/pkg/helpers/tokencmd"
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
 	userclient "github.com/openshift/origin/pkg/user/generated/internalclientset/typed/user/internalversion"
 	testutil "github.com/openshift/origin/test/util"
 	testserver "github.com/openshift/origin/test/util/server"
+	configapi "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
 )
 
 func TestOAuthHTPasswd(t *testing.T) {
