+++ test/integration/project_reqlimit_test.go
@@ -11,7 +11,6 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
 	projectapi "github.com/openshift/origin/pkg/project/apis/project"
 	requestlimit "github.com/openshift/origin/pkg/project/apiserver/admission/apis/requestlimit"
 	projectclient "github.com/openshift/origin/pkg/project/generated/internalclientset"
@@ -19,6 +18,7 @@ import (
 	userclient "github.com/openshift/origin/pkg/user/generated/internalclientset/typed/user/internalversion"
 	testutil "github.com/openshift/origin/test/util"
 	testserver "github.com/openshift/origin/test/util/server"
+	configapi "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
 )
 
 func setupProjectRequestLimitTest(t *testing.T, pluginConfig *requestlimit.ProjectRequestLimitConfig) (kubernetes.Interface, *rest.Config, func()) {
