+++ test/integration/runonce_duration_test.go
@@ -7,9 +7,9 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	pluginapi "github.com/openshift/origin/pkg/autoscaling/admission/apis/runonceduration"
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
 	testutil "github.com/openshift/origin/test/util"
 	testserver "github.com/openshift/origin/test/util/server"
+	configapi "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
 )
 
 func testRunOnceDurationPod(activeDeadlineSeconds int64) *corev1.Pod {
