+++ test/util/server/deprecated_openshift/apis/config/install/install.go
@@ -8,8 +8,8 @@ import (
 	auditv1alpha1 "k8s.io/apiserver/pkg/apis/audit/v1alpha1"
 	auditv1beta1 "k8s.io/apiserver/pkg/apis/audit/v1beta1"
 
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
-	configapiv1 "github.com/openshift/origin/pkg/cmd/server/apis/config/v1"
+	configapi "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
+	configapiv1 "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config/v1"
 )
 
 func init() {
