+++ test/util/server/deprecated_openshift/apis/config/latest/latest.go
@@ -4,7 +4,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
+	configapi "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
 )
 
 // HACK TO ELIMINATE CYCLE UNTIL WE KILL THIS PACKAGE
