+++ test/util/server/deprecated_openshift/apis/config/register.go
