+++ test/util/server/deprecated_openshift/apis/config/stringsource.go
