+++ test/util/server/deprecated_openshift/apis/config/v1/doc.go
@@ -1,0 +1,7 @@
+// +k8s:conversion-gen=github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config
+// +k8s:conversion-gen-external-types=github.com/openshift/api/legacyconfig/v1
+// +k8s:defaulter-gen=TypeMeta
+// +k8s:defaulter-gen-input=../../../../../../../../../../github.com/openshift/api/legacyconfig/v1
+
+// Package v1 is the v1 version of the API.
+package v1
