+++ test/util/server/deprecated_openshift/apis/config/v1/register.go
@@ -8,7 +8,7 @@ import (
 	buildv1 "github.com/openshift/api/build/v1"
 	legacyconfigv1 "github.com/openshift/api/legacyconfig/v1"
 	buildinternalconversions "github.com/openshift/origin/pkg/build/apis/build/v1"
-	"github.com/openshift/origin/pkg/cmd/server/apis/config"
+	"github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
 )
 
 var (
