+++ test/util/server/server.go
@@ -46,10 +46,10 @@ import (
 	openshiftapiserver "github.com/openshift/origin/pkg/cmd/openshift-apiserver"
 	openshiftcontrollermanager "github.com/openshift/origin/pkg/cmd/openshift-controller-manager"
 	openshiftkubeapiserver "github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver"
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
 	newproject "github.com/openshift/origin/pkg/oc/cli/admin/project"
 	"github.com/openshift/origin/test/util"
+	configapi "github.com/openshift/origin/test/util/server/deprecated_openshift/apis/config"
 	"github.com/openshift/origin/test/util/server/deprecated_openshift/configconversion"
 	"github.com/openshift/origin/test/util/server/deprecated_openshift/deprecatedcerts"
 	"github.com/openshift/origin/test/util/server/deprecated_openshift/deprecatedclient"
